# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: JAO LAB Fashion & Styles commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [catalog]
      summary: Search and filter products
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [featured, newest, price-low, price-high]
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [catalog]
      summary: Get a product by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /wishlist/{productId}:
    post:
      operationId: addWishlistItem
      tags: [commerce]
      summary: Add a product to the wishlist
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated wishlist
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    delete:
      operationId: removeWishlistItem
      tags: [commerce]
      summary: Remove a product from the wishlist
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated wishlist
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /orders:
    get:
      operationId: listOrders
      tags: [commerce]
      summary: List current customer's orders
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [commerce]
      summary: Create an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [commerce]
      summary: Get order details and tracking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /profile:
    get:
      operationId: getProfile
      tags: [commerce]
      summary: Get the current customer's profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Customer profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
    patch:
      operationId: updateProfile
      tags: [commerce]
      summary: Update the current customer's profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileInput"
      responses:
        "200":
          description: Updated customer profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /cart:
    get:
      operationId: getCart
      tags: [commerce]
      summary: Get the current customer's cart
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Customer cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    put:
      operationId: replaceCart
      tags: [commerce]
      summary: Replace the current customer's cart
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartInput"
      responses:
        "200":
          description: Updated customer cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: clearCart
      tags: [commerce]
      summary: Clear the current customer's cart
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Empty customer cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Profile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        fullName:
          type: string
        phone:
          type: string
        state:
          type: string
        city:
          type: string
        address:
          type: string
      required: [id, email, fullName, phone, state, city, address]
    ProfileInput:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        state:
          type: string
        city:
          type: string
        address:
          type: string
      required: [fullName, phone, state, city, address]
    CartItemInput:
      type: object
      properties:
        productId:
          type: string
        quantity:
          type: integer
        size:
          type: string
        color:
          type: string
      required: [productId, quantity, size, color]
    CartLine:
      type: object
      properties:
        product:
          $ref: "#/components/schemas/Product"
        quantity:
          type: integer
        size:
          type: string
        color:
          type: string
      required: [product, quantity, size, color]
    Cart:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartLine"
      required: [items]
    CartInput:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItemInput"
      required: [items]
    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
        subcategory:
          type: string
        price:
          type: number
        compareAtPrice:
          type: ["number", "null"]
        image:
          type: string
        images:
          type: array
          items:
            type: string
        sizes:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
        stock:
          type: integer
        tags:
          type: array
          items:
            type: string
        featured:
          type: boolean
        newest:
          type: boolean
        trending:
          type: boolean
        installmentAvailable:
          type: boolean
        installmentAmount:
          type: ["number", "null"]
      required:
        - id
        - name
        - description
        - category
        - subcategory
        - price
        - compareAtPrice
        - image
        - images
        - sizes
        - colors
        - stock
        - tags
        - featured
        - newest
        - trending
        - installmentAvailable
        - installmentAmount
...
    OrderItem:
      type: object
      properties:
        productId:
          type: string
        productName:
          type: string
        image:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        size:
          type: string
        color:
          type: string
      required: [productId, productName, image, quantity, unitPrice, size, color]
    Order:
      type: object
      properties:
        id:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        total:
          type: number
        paid:
          type: number
        remaining:
          type: number
        paymentMethod:
          type: string
          enum: [full, installment, bank-transfer]
        installmentFrequency:
          type: ["string", "null"]
          enum: [weekly, monthly, null]
        status:
          type: string
          enum: [pending, payment-verification, confirmed, processing, shipped, out-for-delivery, delivered, cancelled]
        createdAt:
          type: string
        expectedDelivery:
          type: string
        address:
          type: string
        nextPayment:
          type: ["string", "null"]
        paymentHistory:
          type: array
          items:
            $ref: "#/components/schemas/PaymentHistoryItem"
        timeline:
          type: array
          items:
            $ref: "#/components/schemas/OrderTimelineItem"
      required: [id, items, total, paid, remaining, paymentMethod, installmentFrequency, status, createdAt, expectedDelivery, address, nextPayment, paymentHistory, timeline]
    PaymentHistoryItem:
      type: object
      properties:
        amount:
          type: number
        date:
          type: string
        status:
          type: string
        reference:
          type: string
      required: [amount, date, status, reference]
    OrderTimelineItem:
      type: object
      properties:
        label:
          type: string
        date:
          type: ["string", "null"]
        complete:
          type: boolean
        current:
          type: boolean
      required: [label, date, complete, current]
    OrderInput:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        fullName:
          type: string
        phone:
          type: string
        email:
          type: string
        state:
          type: string
        city:
          type: string
        address:
          type: string
        notes:
          type: string
        paymentMethod:
          type: string
          enum: [full, installment, bank-transfer]
        installmentFrequency:
          type: ["string", "null"]
          enum: [weekly, monthly, null]
      required: [items, fullName, phone, email, state, city, address, notes, paymentMethod]
...