# lib/api-zod/src/generated/types/cartInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * JAO LAB Fashion & Styles commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { CartItemInput } from './cartItemInput';

export interface CartInput {
  items: CartItemInput[];
}