# lib/api-zod/src/generated/types/cartItemInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * JAO LAB Fashion & Styles commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface CartItemInput {
  productId: string;
  quantity: number;
  size: string;
  color: string;
}