# lib/api-zod/src/generated/types/cartLine.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * JAO LAB Fashion & Styles commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from './product';

export interface CartLine {
  product: Product;
  quantity: number;
  size: string;
  color: string;
}