# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * JAO LAB Fashion & Styles commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderInputInstallmentFrequency } from './orderInputInstallmentFrequency';
import type { OrderInputPaymentMethod } from './orderInputPaymentMethod';
import type { OrderItem } from './orderItem';

export interface OrderInput {
  items: OrderItem[];
  fullName: string;
  phone: string;
  email: string;
  state: string;
  city: string;
  address: string;
  notes: string;
  paymentMethod: OrderInputPaymentMethod;
  /** @nullable */
  installmentFrequency: OrderInputInstallmentFrequency;
}